Moved UsersService under App\Services namespace
