+++ code/app/Services/UsersService.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace App;
+namespace App\Services;
 
 use App\Exceptions\AuthException;
+use App\User;
 use Auth;
 use DB;
 use Hash;
